backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
aiofiles==23.2.1
requests==2.31.0
dnspython==2.4.2
python-whois==0.9.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.105.0
Pillow==10.1.0
